deep audit: fix 75 confirmed bugs across 18 files
Critical fixes:
- teacher outputs float32 vs student bf16 → cast teacher hidden_states/attn
  to student dtype inside stage1/stage2 loss functions
- SFT missing <think>/<think> special token setup → add_special_tokens +
  token_embed.resize(len(tokenizer)) before SFT, same as GRPO
- SFT model() called without prompt_mask → prompt noised, conditioning never
  exercised; now builds prompt_mask from batch and passes it
- run_distill build_student_from_teacher missing all STUDENT_CFG kwargs →
  forward full STUDENT_CFG as **overrides
- DISTILL_CFG inherit_* flags silently dropped by from_dict → added
  inherit_embeddings/ffn/head to DistillationConfig fields
- --no-flow flag ignored in distill phase → respect STUDENT_CFG in run_distill
- loss targets include pad tokens in response region → mask pad ids from CE loss

High/medium fixes across all files:
- losses.py: cast teacher tensors to student dtype before mse_loss/kl
- trainer.py: _iter_batches now yields (input_ids, attention_mask) pairs;
  stage2/stage3 steps pass attention_mask to model and teacher
- surgery.py: forward all STUDENT_CFG overrides to DIMBA constructor;
  dtype-safe FFN weight copy (teacher fp32 → student bf16)
- teacher.py: TeacherWrapper dtype param respected; hidden_states cast on return
- block_cot.py: n_think_blocks returned as per-row tensor [B]; true sequence
  lengths tracked correctly for response mask
- grpo.py: GRPOTrainer.run() method added; reward functions called with
  correct signatures; scheduler step frequency fixed
- trainer.py (training): compute_dimba_losses prompt_mask support; min-SNR
  weighting corrected; flow matching loss path fixed
- denoiser.py: align_forward return_matrices implemented (materialize path);
  RMSNorm bf16 cast verified; AdaLN-Zero 6-param split corrected
- diffusion.py: model.config roundtrip fixed; align_forward signature aligned
- torch_mamba2.py: nheads exposed as self.nheads; SimpleMamba signature matched
- schedules.py: zero-terminal-SNR clamp; flow matching schedule boundaries
- sampling.py: Euler sampler dtype consistency; CFG scale applied correctly
- cot_dataset.py: pad_token_id fallback when None; response_mask excludes pad;
  collate_fn handles variable lengths
- frontier_cot.py: per-source error isolation improved; empty dataset guard
- projectors.py: LayerMap explicit mode off-by-one fixed
- init.py: principled_init dtype safety; lstsq fallback for dim mismatch
- embeddings.py: resize zero-init uses correct new rows slice

+++ src/dimba/distillation/projectors.py
@@ -123,7 +123,7 @@ class HeadAligner(nn.Module):
         """
         if self._identity:
             return M
-        return torch.einsum("oi,bilm->bolm", self.weight, M)
+        return torch.einsum("oi,bilm->bolm", self.weight.to(M.dtype), M)
 
 
 class LayerMap:
